Catch PHP exception thrown during callback execution
Apart from a result, PHP callback can throw an excetion. Thrown exception
is stored it global variable "exception" until it's caught. Once this variable
is set, PHP internal methods stop working considering they are in an
stack-unwinding process. As soon as lua extension doesn't cleanup this state,
all following callbacks return nil until the exceptoin is caught and the
state is cleaned up.

However, the exception is caught only when lua code is done and execution flow is
returned to PHP again.